change data type to size_t for iteration

+++ 0x08-recursion/0-puts_recursion.c
@@ -8,7 +8,7 @@
 
 void _puts_recursion(char *s)
 {
-	long int i;
+	size_t i;
 
 	for (i = 0; i <= strlen(s); i++)
 	{
